Cleaned code, and Moved the files into models to follow standard structure.
