+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/controller/ImageController.java
@@ -1,6 +1,6 @@
 package com.parkrangers.parkquest_backend.controller;
 
-import com.parkrangers.parkquest_backend.dto.ImageDTO;
+import com.parkrangers.parkquest_backend.models.dto.ImageDTO;
 import com.parkrangers.parkquest_backend.service.ParkService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
