--- parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/controllers/GoogleAuthController.java
@@ -1,4 +1,0 @@
-package com.parkrangers.parkquest_backend.controllers;
-
-public class GoogleAuthController {
-}
