--- parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/dto/ActivityDTO.java
@@ -1,4 +1,0 @@
-package com.parkrangers.parkquest_backend.dto;
-
-public class ActivityDTO {
-}
