--- parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/models/Role.java
@@ -1,4 +1,0 @@
-package com.parkrangers.parkquest_backend.models;
-
-public class Role {
-}
