--- parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/models/User.java
@@ -1,4 +1,0 @@
-package com.parkrangers.parkquest_backend.models;
-
-public class User {
-}
