+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/models/dto/ActivityDTO.java
@@ -1,0 +1,4 @@
+package com.parkrangers.parkquest_backend.models.dto;
+
+public class ActivityDTO {
+}
