+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/models/entity/Role.java
@@ -1,0 +1,4 @@
+package com.parkrangers.parkquest_backend.models.entity;
+
+public class Role {
+}
