+++ parkquest-backend/src/main/java/com/parkrangers/parkquest_backend/models/entity/TripPlan.java
@@ -1,4 +1,4 @@
-package com.parkrangers.parkquest_backend.models;
+package com.parkrangers.parkquest_backend.models.entity;
 
 
 import com.parkrangers.parkquest_backend.models.response.Park;
